Don't unsafely read buffer (#40)
rustls recently added new plaintext buffer access functions that allow for safely appending to a tokio ReadBuf.

The original implementation was technically not sound, though it never caused any issues, because it transmuted a MaybeUninit buffer to a &mut [u8]. Unfortunately at the time, this was the only real way to read data in a performant way. This could potentially cause miscompilations in certain cases, though I never personally saw that occur.

Instead of doing the unsafe thing, we bump rustls to the latest version and use the new fill_buf method along with put_slice. To ensure we don't lose performance, we will read as much plaintext as we can each time through the try_read loop.

A new debug assertion ensures we return the correct number of bytes read.